+++ packages/xmc/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sitecore-marketplace-sdk/xmc
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [3b308a9]
+- Updated dependencies [0a7363e]
+  - @sitecore-marketplace-sdk/client@0.3.0
+
+## 0.3.0
+
+### Minor Changes
+
+- 56c82ec: Added Agent API documentation
+- d936872: Added new Agent API in the xmc module
+
 ## 0.2.2
 
 ### Patch Changes